// src/contracts/CandidateProfile/interfaces.ts
export interface CandidateProfile {
    key: string;
    name: string;
    aggregatedScore: number;
    email: string;
    phone: string;
    currentPosition: string;
    currentCompany: string;
    totalExperience: number;
    employmentStartDate: string;
    experienceWithCurrentCompany: number;
    location: string;
    skills: string[];
    certifications: string[];
    education: string;
    languages: string[];
    linkedinProfile: string;
    githubProfile: string;
    portfolio: string;
    availability: string;
    expectedSalary: string;
    references: string;
    workAuthorization: string;
    projects: string[];
    summary: string;
    lastUpdated: string;
  }
  